requirements.txt
allure-pytest
requests
pytest
api
